+++ components/emulators/common/src/CMakeLists.txt
@@ -21,6 +21,11 @@ add_library(emulator_common
 
 target_compile_features(emulator_common PUBLIC cxx_std_17)
 
+# Place generated Fortran .mod files in a predictable directory so that
+# downstream Fortran targets (e.g. the driver test) can find them.
+set_target_properties(emulator_common PROPERTIES
+  Fortran_MODULE_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/modules)
+
 target_include_directories(emulator_common PUBLIC
   ${CMAKE_CURRENT_SOURCE_DIR}
   ${CMAKE_CURRENT_SOURCE_DIR}/inference
@@ -28,5 +33,6 @@ target_include_directories(emulator_common PUBLIC
 
 target_include_directories(emulator_common PUBLIC
   ${CMAKE_CURRENT_BINARY_DIR}
+  ${CMAKE_CURRENT_BINARY_DIR}/modules
 )
